+++ core/utility/strings.cpp
@@ -12,7 +12,8 @@ namespace strings
 		const auto len = _vscwprintf(szMsg, argList);
 		if (0 != len)
 		{
-			auto buffer = std::wstring(len + 1, '\0'); // Include extra since _vsnwprintf_s writes a null terminator
+			auto buffer =
+				std::wstring((size_t) len + 1, '\0'); // Include extra since _vsnwprintf_s writes a null terminator
 			if (_vsnwprintf_s(const_cast<wchar_t*>(buffer.c_str()), len + 1, _TRUNCATE, szMsg, argList) > 0)
 			{
 				buffer.resize(len); // Resize to exact length before returning
